Fix static code issues
